pwm: pwm-dwc: Change duty and period calculation variables to 64-bit
HW can have longer duty and period than the PWM core can express. The
high and low period counters are 32-bit running at 10 ns clock period so
HW can do maximum ~42.9 s duty cycle with ~85.9 s period while PWM core
can express maximum ~4.29 s.

Change the variables here to 64-bit in order to avoid overflow in
calculation, to be able to print a warning and to be prepared if PWM
core changes to 64-bit integers in the future.

+++ drivers/pwm/pwm-dwc.c
@@ -140,7 +140,7 @@ static void dwc_pwm_get_state(struct pwm_chip *chip, struct pwm_device *pwm,
 			      struct pwm_state *state)
 {
 	struct dwc_pwm *dwc = to_dwc_pwm(chip);
-	u32 duty, period;
+	u64 duty, period;
 
 	pm_runtime_get_sync(dwc->dev);
